ai-evaluation-dataset/requirements.txt
# Install with: pip install -r requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
SPARQLWrapper>=2.0.0
PyPDF2>=3.0.1
